+++ src/PowerShell/Commands/GetPartnerProduct.cs
@@ -155,15 +155,11 @@ namespace Microsoft.Store.PartnerCenter.PowerShell.Commands
             targetView.AssertNotEmpty(nameof(targetView));
             targetSegment.AssertNotEmpty(nameof(targetSegment));
 
-            try
+            products = Partner.Products.ByCountry(countryCode).ByTargetView(targetView).ByTargetSegment(targetSegment).GetAsync().GetAwaiter().GetResult();
+
+            if (products.TotalCount > 0)
             {
-                products = Partner.Products.ByCountry(countryCode).ByTargetView(targetView).ByTargetSegment(targetSegment).GetAsync().GetAwaiter().GetResult();
-                if (products.TotalCount > 0)
-                    WriteObject(products.Items.Select(p => new PSProduct(p)), true);
-            }
-            catch (PartnerCenter.Exceptions.PartnerException ex)
-            {
-                throw new PSPartnerException("Error getting products for segment: " + targetSegment, ex);
+                WriteObject(products.Items.Select(p => new PSProduct(p)), true);
             }
         }
     }