+++ ChangeLog.md
@@ -22,11 +22,22 @@
 
 ## Upcoming Release
 
-* Corrected issue #90
-    - This issue was causing user directory role commands were not functioning as expected.
+* Auditing
+  * Added new operation and resource types
+* Customers
+  * Added the ability to manage customer qualifications
+  * Corrected issue #90 that was causing directory role operations to not function as excepted
+* Entitlements
+  * Added the AlternateId property to the reference order object
+* Invoice
+  * Added the ability to download the Taiwan tax receipt
+* Orders
+  * Addressed an issue with requesting an order by the billing cycle
+* Products
+  * Corrected an issue with requesting products by country, target view, and target segment
 
 ## Version 1.5.1902.5
 
-* Added the New-PartnerCustomerApplicationConsent command 
-    - This command can be used to create a new application consent for the specified customer.
+* Added the New-PartnerCustomerApplicationConsent command
+  * This command can be used to create a new application consent for the specified customer
 * Update to version 10.0.3 of Newtonsoft.Json